requirements.txt
fastapi==0.116.1
openai==1.107.2
pydantic==2.11.7
python-dotenv==1.1.1
pytest==8.4.1
uvicorn==0.35.0